FS2: Add Process1Spec that tests the current transducers
